+++ src/sdk/shared/interfaces/Component.ts
@@ -1,0 +1,4 @@
+export interface Component {
+  purl: string;
+  requirement?: string;
+}