+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [0.18.0] (2025-08-28)
+### Added 
+- Added REST protocol implementation for SCANOSS dependency service
+- Added proxy options to `dep` subcommand
+### Fixed
+- Improved URL path handling for free vs premium SCANOSS endpoints
+
 ## [0.17.3] (2025-06-27)
 ### Fixed
 - Empty winnowing.wfp file now created when scanning empty directories or filtered content
@@ -101,3 +108,5 @@ All notable changes to this project will be documented in this file. See [standa
 ### [0.17.0](https://github.com/scanoss/scanoss.js/compare/v0.16.5...v0.17.0) (2025-06-10)
 ### [0.17.1](https://github.com/scanoss/scanoss.js/compare/v0.17.0...v0.17.1) (2025-06-16)
 ### [0.17.2](https://github.com/scanoss/scanoss.js/compare/v0.17.1...v0.17.2) (2025-06-25)
+### [0.17.3](https://github.com/scanoss/scanoss.js/compare/v0.17.2...v0.17.3) (2025-06-27)
+### [0.17.4](https://github.com/scanoss/scanoss.js/compare/v0.17.3...v0.17.4) (2025-08-27)